pyproject.toml
[project]
name = "ai-newsletter"
version = "0.1.0"
description = "Generate a recurring topic-specific newsletter from a reusable skill file"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.11",
    "google-genai>=1.7.0",
    "markdown>=3.7",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "streamlit>=1.43.0",
    "trafilatura>=1.12.2",
]

[tool.uv]
package = false
